Rename module and reformat code
