+++ src/bidimensional/__init__.py
@@ -1,3 +1,3 @@
-from .core.coordinates import Coordinate
+from .core.coordinate import Coordinate
 from .core.lines import Line, Segment
 from .core import operations
